Web page with Movie Search and Wishlist
In this project, a web page integrated with an API was developed, which searches and displays information about films. When searching for a movie, the data is presented in a modal, including the image and detailed movie information. Additionally, you can save the movie to a wish list to watch later.